requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
streamlit>=1.0.0
plotly>=5.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0
requests>=2.26.0
python-dotenv>=0.19.0
geopy>=2.3.0
